New XioConnection management logic.
Introduce a new connection list, protected by conns_sp.  Residence
on the conns_list connotes a refcnt of +1, the sentinel ref, and all
new connections (ACTIVE or PASSIVE), are enqueued on creation.

In addition, we now add PASSIVE connections to conns_entity_map when
when the actual peer_addr is known.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>