Move delivery receipt to trailer.
The delivery receipt was arriving with the first message
delivery because we configured it to do so, and it's not
legal to destruct an entire sequence after delivery of the
first (hence the original multiple refs).

On the other hand, let's try a single ref, with the delivery
receipt on the last in sequence.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>