Another (hopefully temporary) msgref change.
It now appears that we MUST currently request delivery
receipts on all messages in a chain.

It also appears that a reliable algorithm for disposing
sender-side messages without delivery receipts (at least
in the common case) is necessary to get full performance
with one-way.  We're still trying to convince Eyal of
that.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>