Remove need for std::list<xio_msg*> (optimization).
For incoming messages, the Accelio xio_msg user_context field
was unused (in an outgoing message, it points to the corresponding
XioMsg).

This change replaces the explicit lists of message pointers that
were being used in XioConnection (its current incoming message
sequence) and XioCompletionHook (tracking the set of constituent
xio_msg structures until reclaim) with a C-style simple list. The
new XioInSeq class is a wrapper for the head and tail pointers,
plus housekeeping data.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>