cls_rgw: fetch more entries until bucket listing done
Fixes: #12913

When listing bucket, we need to read more bucket index entries,
until either we got enough entries, or we're passed the index
namespace. The problem was that with versioning we end up
skipping certain entries, so we may end up returning an empty
result and won't update the listing marker.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>