lua: build Lua-5.3.2 as library
Embed the Lua VM into Ceph as a convenience library. Rather than depend
on an external platform specific dependency this allows us to have full
control over the VM (e.g. modifying it to suit Ceph sandbox needs).

Currently this is linked into cls_lua, but could be used in other parts
of Ceph such as the VM used for the Mantle MDS balancer.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>