osd: probabilistic promotion throttling
Global tunables osd_tier_promote_max_{objects,bytes}_sec adjust a
probability that any given promotion we might otherwise do will be done
or rejected.  Based on the observed rate of promotion, and the target/max,
we will adjust the probability up for down.

This avoids the complexity of a throttle queue and gets us a reasonably
accurate throttle.  What it doesn't yet do is cope with thundering herds of
promotions.

Signed-off-by: Sage Weil <sage@redhat.com>