ceph-disk: Always set user/group when running ceph-osd
Currently, we are running ceph-osd daemons from ceph-disk without
supplying it with the setuser/group arguments. This causes issues
because we are accessing files that are owned by ceph user with root
which is not explicitely allowed in SELinux scenario.

Simply pass the setuser/group arguments to all the occurrences of
ceph-osd call to fix this.

Fixes: #14870
Signed-off-by: Boris Ranto <branto@redhat.com>