global: Drop privileges as soon as possible
Currently, we drop privileges after we read a config file. This results
in dac_overrid denials because we access ceph owned files with root user
which is not allowed by default in SELinux environment.

This patch moves the logic that drops the privileges to an earlier phase
in global_pre_init function before the ceph daemon config file is read
and accessed.

Fixes: #14870
Signed-off-by: Boris Ranto <branto@redhat.com>